requirements.txt
connexion[swagger-ui]
coloredlogs
pyaml
